> moved commandData to command handler dir closes #23
